Move command line commands to iggy::cmd module
Move modules with structures which implement command line commands
to separate module iggy::cmd in order to keep it separate from other
iggy SDK code and make it optional using iggy-cmd feature in iggy
crate. Feature is being used only in iggy-cmd crate.
