+++ cmd/Cargo.toml
@@ -1,12 +1,12 @@
 [package]
 name = "iggy-cmd"
-version = "0.4.0"
+version = "0.4.1"
 edition = "2021"
 authors = ["bartosz.ciesla@gmail.com"]
 repository = "https://github.com/iggy-rs/iggy"
 
 [dependencies]
-iggy = { path = "../iggy", default-features = false }
+iggy = { path = "../iggy", default-features = false, features = ["iggy-cmd"] }
 clap = { version = "4.1.11", features = ["derive"] }
 tokio = { version = "1.28.2", features = ["full"] }
 async-trait = "0.1.68"
