+++ iggy/src/cmd/mod.rs
@@ -1,0 +1,3 @@
+pub mod partitions;
+pub mod streams;
+pub mod topics;
