+++ iggy/src/cmd/partitions/mod.rs
@@ -1,0 +1,2 @@
+pub mod create_partitions;
+pub mod delete_partitions;
