+++ iggy/src/streams/delete_stream.rs
@@ -1,16 +1,11 @@
-use crate::cli_command::CliCommand;
-use crate::client::Client;
+use crate::bytes_serializable::BytesSerializable;
 use crate::command::CommandPayload;
 use crate::error::Error;
 use crate::identifier::Identifier;
 use crate::validatable::Validatable;
-use crate::{bytes_serializable::BytesSerializable, cli_command::PRINT_TARGET};
-use anyhow::Context;
-use async_trait::async_trait;
 use serde::{Deserialize, Serialize};
 use std::fmt::Display;
 use std::str::FromStr;
-use tracing::{event, Level};
 
 #[derive(Debug, Serialize, Deserialize, PartialEq, Default)]
 pub struct DeleteStream {
@@ -61,41 +56,6 @@ impl BytesSerializable for DeleteStream {
 impl Display for DeleteStream {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         write!(f, "{}", self.stream_id)
-    }
-}
-
-pub struct DeleteStreamCmd {
-    delete_stream: DeleteStream,
-}
-
-impl DeleteStreamCmd {
-    pub fn new(stream_id: Identifier) -> Self {
-        Self {
-            delete_stream: DeleteStream { stream_id },
-        }
-    }
-}
-
-#[async_trait]
-impl CliCommand for DeleteStreamCmd {
-    fn explain(&self) -> String {
-        format!("delete stream with ID: {}", self.delete_stream.stream_id)
-    }
-
-    async fn execute_cmd(&mut self, client: &dyn Client) -> anyhow::Result<(), anyhow::Error> {
-        client
-            .delete_stream(&self.delete_stream)
-            .await
-            .with_context(|| {
-                format!(
-                    "Problem deleting stream with ID: {}",
-                    self.delete_stream.stream_id
-                )
-            })?;
-
-        event!(target: PRINT_TARGET, Level::INFO, "Stream with ID: {} deleted", self.delete_stream.stream_id);
-
-        Ok(())
     }
 }
