+++ iggy/src/streams/get_streams.rs
@@ -1,17 +1,10 @@
 use crate::bytes_serializable::BytesSerializable;
-use crate::cli_command::{CliCommand, PRINT_TARGET};
-use crate::client::Client;
 use crate::command::CommandPayload;
 use crate::error::Error;
-use crate::utils::timestamp::TimeStamp;
 use crate::validatable::Validatable;
-use anyhow::Context;
-use async_trait::async_trait;
-use comfy_table::Table;
 use serde::{Deserialize, Serialize};
 use std::fmt::Display;
 use std::str::FromStr;
-use tracing::{event, Level};
 
 #[derive(Debug, Default, Serialize, Deserialize, PartialEq)]
 pub struct GetStreams {}
@@ -56,95 +49,6 @@ impl BytesSerializable for GetStreams {
 impl Display for GetStreams {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         write!(f, "")
-    }
-}
-
-pub enum GetStreamsOutput {
-    Table,
-    List,
-}
-
-pub struct GetStreamsCmd {
-    get_streams: GetStreams,
-    output: GetStreamsOutput,
-}
-
-impl GetStreamsCmd {
-    pub fn new(output: GetStreamsOutput) -> Self {
-        GetStreamsCmd {
-            get_streams: GetStreams {},
-            output,
-        }
-    }
-}
-
-impl Default for GetStreamsCmd {
-    fn default() -> Self {
-        GetStreamsCmd {
-            get_streams: GetStreams {},
-            output: GetStreamsOutput::Table,
-        }
-    }
-}
-
-#[async_trait]
-impl CliCommand for GetStreamsCmd {
-    fn explain(&self) -> String {
-        let mode = match self.output {
-            GetStreamsOutput::Table => "table",
-            GetStreamsOutput::List => "list",
-        };
-        format!("list streams in {mode} mode")
-    }
-
-    async fn execute_cmd(&mut self, client: &dyn Client) -> anyhow::Result<(), anyhow::Error> {
-        let streams = client
-            .get_streams(&self.get_streams)
-            .await
-            .with_context(|| String::from("Problem getting list of streams"))?;
-
-        if streams.is_empty() {
-            event!(target: PRINT_TARGET, Level::INFO, "No streams found!");
-            return Ok(());
-        }
-
-        match self.output {
-            GetStreamsOutput::Table => {
-                let mut table = Table::new();
-
-                table.set_header(vec![
-                    "ID", "Created", "Name", "Size (B)", "Messages", "Topics",
-                ]);
-
-                streams.iter().for_each(|stream| {
-                    table.add_row(vec![
-                        format!("{}", stream.id),
-                        TimeStamp::from(stream.created_at).to_string("%Y-%m-%d %H:%M:%S"),
-                        stream.name.clone(),
-                        format!("{}", stream.size_bytes),
-                        format!("{}", stream.messages_count),
-                        format!("{}", stream.topics_count),
-                    ]);
-                });
-
-                event!(target: PRINT_TARGET, Level::INFO, "{table}");
-            }
-            GetStreamsOutput::List => {
-                streams.iter().for_each(|stream| {
-                    event!(target: PRINT_TARGET, Level::INFO,
-                        "{}|{}|{}|{}|{}|{}",
-                        stream.id,
-                        TimeStamp::from(stream.created_at).to_string("%Y-%m-%d %H:%M:%S"),
-                        stream.name,
-                        stream.size_bytes,
-                        stream.messages_count,
-                        stream.topics_count
-                    );
-                });
-            }
-        }
-
-        Ok(())
     }
 }
