+++ iggy/src/streams/update_stream.rs
@@ -1,18 +1,13 @@
 use crate::bytes_serializable::BytesSerializable;
-use crate::cli_command::{CliCommand, PRINT_TARGET};
-use crate::client::Client;
 use crate::command::CommandPayload;
 use crate::error::Error;
 use crate::identifier::Identifier;
 use crate::utils::text;
 use crate::validatable::Validatable;
-use anyhow::Context;
-use async_trait::async_trait;
 use bytes::BufMut;
 use serde::{Deserialize, Serialize};
 use std::fmt::Display;
 use std::str::{from_utf8, FromStr};
-use tracing::{event, Level};
 
 const MAX_NAME_LENGTH: usize = 255;
 
@@ -99,47 +94,6 @@ impl BytesSerializable for UpdateStream {
 impl Display for UpdateStream {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         write!(f, "{}|{}", self.stream_id, self.name)
-    }
-}
-
-pub struct UpdateStreamCmd {
-    update_stream: UpdateStream,
-}
-
-impl UpdateStreamCmd {
-    pub fn new(stream_id: Identifier, name: String) -> Self {
-        UpdateStreamCmd {
-            update_stream: UpdateStream { stream_id, name },
-        }
-    }
-}
-
-#[async_trait]
-impl CliCommand for UpdateStreamCmd {
-    fn explain(&self) -> String {
-        format!(
-            "update stream with ID: {} and name: {}",
-            self.update_stream.stream_id, self.update_stream.name
-        )
-    }
-
-    async fn execute_cmd(&mut self, client: &dyn Client) -> anyhow::Result<(), anyhow::Error> {
-        client
-            .update_stream(&self.update_stream)
-            .await
-            .with_context(|| {
-                format!(
-                    "Problem updating stream with ID: {} and name: {}",
-                    self.update_stream.stream_id, self.update_stream.name
-                )
-            })?;
-
-        event!(target: PRINT_TARGET, Level::INFO,
-            "Stream with ID: {} updated name: {} ",
-            self.update_stream.stream_id, self.update_stream.name
-        );
-
-        Ok(())
     }
 }
