+++ iggy/src/topics/delete_topic.rs
@@ -1,16 +1,11 @@
 use crate::bytes_serializable::BytesSerializable;
-use crate::cli_command::{CliCommand, PRINT_TARGET};
-use crate::client::Client;
 use crate::command::CommandPayload;
 use crate::error::Error;
 use crate::identifier::Identifier;
 use crate::validatable::Validatable;
-use anyhow::Context;
-use async_trait::async_trait;
 use serde::{Deserialize, Serialize};
 use std::fmt::Display;
 use std::str::FromStr;
-use tracing::{event, Level};
 
 #[derive(Debug, Serialize, Deserialize, PartialEq, Default)]
 pub struct DeleteTopic {
@@ -78,50 +73,6 @@ impl BytesSerializable for DeleteTopic {
 impl Display for DeleteTopic {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         write!(f, "{}|{}", self.stream_id, self.topic_id)
-    }
-}
-
-pub struct DeleteTopicCmd {
-    delete_topic: DeleteTopic,
-}
-
-impl DeleteTopicCmd {
-    pub fn new(stream_id: Identifier, topic_id: Identifier) -> Self {
-        Self {
-            delete_topic: DeleteTopic {
-                stream_id,
-                topic_id,
-            },
-        }
-    }
-}
-
-#[async_trait]
-impl CliCommand for DeleteTopicCmd {
-    fn explain(&self) -> String {
-        format!(
-            "delete topic with ID: {} in stream with ID: {}",
-            self.delete_topic.topic_id, self.delete_topic.stream_id
-        )
-    }
-
-    async fn execute_cmd(&mut self, client: &dyn Client) -> anyhow::Result<(), anyhow::Error> {
-        client
-            .delete_topic(&self.delete_topic)
-            .await
-            .with_context(|| {
-                format!(
-                    "Problem deleting topic with ID: {} in stream {}",
-                    self.delete_topic.topic_id, self.delete_topic.stream_id
-                )
-            })?;
-
-        event!(target: PRINT_TARGET, Level::INFO,
-            "Topic with ID: {} in stream with ID: {} deleted",
-            self.delete_topic.topic_id, self.delete_topic.stream_id
-        );
-
-        Ok(())
     }
 }
