+++ iggy/src/topics/get_topics.rs
@@ -1,18 +1,11 @@
 use crate::bytes_serializable::BytesSerializable;
-use crate::cli_command::{CliCommand, PRINT_TARGET};
-use crate::client::Client;
 use crate::command::CommandPayload;
 use crate::error::Error;
 use crate::identifier::Identifier;
-use crate::utils::timestamp::TimeStamp;
 use crate::validatable::Validatable;
-use anyhow::Context;
-use async_trait::async_trait;
-use comfy_table::Table;
 use serde::{Deserialize, Serialize};
 use std::fmt::Display;
 use std::str::FromStr;
-use tracing::{event, Level};
 
 #[derive(Debug, Serialize, Deserialize, PartialEq, Default)]
 pub struct GetTopics {
@@ -63,96 +56,6 @@ impl BytesSerializable for GetTopics {
 impl Display for GetTopics {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         write!(f, "{}", self.stream_id)
-    }
-}
-
-pub enum GetTopicsOutput {
-    Table,
-    List,
-}
-
-pub struct GetTopicsCmd {
-    get_topics: GetTopics,
-    output: GetTopicsOutput,
-}
-
-impl GetTopicsCmd {
-    pub fn new(stream_id: Identifier, output: GetTopicsOutput) -> Self {
-        Self {
-            get_topics: GetTopics { stream_id },
-            output,
-        }
-    }
-}
-
-#[async_trait]
-impl CliCommand for GetTopicsCmd {
-    fn explain(&self) -> String {
-        format!(
-            "get topics from stream with ID: {}",
-            self.get_topics.stream_id
-        )
-    }
-
-    async fn execute_cmd(&mut self, client: &dyn Client) -> anyhow::Result<(), anyhow::Error> {
-        let topics = client.get_topics(&self.get_topics).await.with_context(|| {
-            format!(
-                "Problem getting topics from stream {}",
-                self.get_topics.stream_id
-            )
-        })?;
-
-        match self.output {
-            GetTopicsOutput::Table => {
-                let mut table = Table::new();
-
-                table.set_header(vec![
-                    "ID",
-                    "Created",
-                    "Name",
-                    "Size (B)",
-                    "Message Expiry (s)",
-                    "Messages Count",
-                    "Partitions Count",
-                ]);
-
-                topics.iter().for_each(|topic| {
-                    table.add_row(vec![
-                        format!("{}", topic.id),
-                        TimeStamp::from(topic.created_at).to_string("%Y-%m-%d %H:%M:%S"),
-                        topic.name.clone(),
-                        format!("{}", topic.size_bytes),
-                        match topic.message_expiry {
-                            Some(value) => format!("{}", value),
-                            None => String::from("None"),
-                        },
-                        format!("{}", topic.messages_count),
-                        format!("{}", topic.partitions_count),
-                    ]);
-                });
-
-                event!(target: PRINT_TARGET, Level::INFO, "{table}");
-            }
-            GetTopicsOutput::List => {
-                topics.iter().for_each(|topic| {
-                    event!(target: PRINT_TARGET, Level::INFO,
-                        "{}|{}|{}|{}|{}|{}|{}",
-                        topic.id,
-                        TimeStamp::from(topic.created_at).to_string("%Y-%m-%d %H:%M:%S"),
-                        topic.name,
-                        topic.size_bytes,
-                        match topic.message_expiry {
-                            Some(value) => format!("{}", value),
-                            None => String::from("None"),
-                        },
-                        topic.messages_count,
-                        topic.partitions_count
-                    );
-                });
-            }
-        }
-
-        Ok(())
     }
 }
