+++ iggy/src/topics/update_topic.rs
@@ -1,19 +1,13 @@
 use crate::bytes_serializable::BytesSerializable;
-use crate::cli_command::{CliCommand, PRINT_TARGET};
-use crate::client::Client;
 use crate::command::CommandPayload;
 use crate::error::Error;
 use crate::identifier::Identifier;
-use crate::utils::message_expire::MessageExpiry;
 use crate::utils::text;
 use crate::validatable::Validatable;
-use anyhow::Context;
-use async_trait::async_trait;
 use bytes::BufMut;
 use serde::{Deserialize, Serialize};
 use std::fmt::Display;
 use std::str::{from_utf8, FromStr};
-use tracing::{event, Level};
 
 const MAX_NAME_LENGTH: usize = 255;
 
@@ -144,81 +138,6 @@ impl Display for UpdateTopic {
             self.message_expiry.unwrap_or(0),
             self.name
         )
-    }
-}
-
-pub struct UpdateTopicCmd {
-    update_topic: UpdateTopic,
-    message_expiry: Option<MessageExpiry>,
-}
-
-impl UpdateTopicCmd {
-    pub fn new(
-        stream_id: Identifier,
-        topic_id: Identifier,
-        name: String,
-        message_expiry: Option<MessageExpiry>,
-    ) -> Self {
-        Self {
-            update_topic: UpdateTopic {
-                stream_id,
-                topic_id,
-                name,
-                message_expiry: match &message_expiry {
-                    None => None,
-                    Some(value) => value.into(),
-                },
-            },
-            message_expiry,
-        }
-    }
-}
-
-#[async_trait]
-impl CliCommand for UpdateTopicCmd {
-    fn explain(&self) -> String {
-        let expiry_text = match &self.message_expiry {
-            Some(value) => format!(" and message expire time: {}", value),
-            None => String::from(""),
-        };
-        format!(
-            "update topic with ID: {}, name: {}{} in stream with ID: {}",
-            self.update_topic.topic_id,
-            self.update_topic.name,
-            expiry_text,
-            self.update_topic.stream_id
-        )
-    }
-
-    async fn execute_cmd(&mut self, client: &dyn Client) -> anyhow::Result<(), anyhow::Error> {
-        client
-            .update_topic(&self.update_topic)
-            .await
-            .with_context(|| {
-                format!(
-                    "Problem updating topic (ID: {}, name: {}{}) in stream with ID: {}",
-                    self.update_topic.topic_id,
-                    self.update_topic.name,
-                    match &self.message_expiry {
-                        Some(value) => format!(" and message expire time: {}", value),
-                        None => String::from(""),
-                    },
-                    self.update_topic.stream_id
-                )
-            })?;
-
-        event!(target: PRINT_TARGET, Level::INFO,
-            "Topic with ID: {} updated name: {}{} in stream with ID: {}",
-            self.update_topic.topic_id,
-            self.update_topic.name,
-            match &self.message_expiry {
-                Some(value) => format!(" and message expire time: {}", value),
-                None => String::from(""),
-            },
-            self.update_topic.stream_id,
-        );
-
-        Ok(())
     }
 }
