+++ iggy/src/utils/mod.rs
@@ -1,5 +1,6 @@
 pub mod checksum;
 pub mod crypto;
+#[cfg(feature = "iggy-cmd")]
 pub mod message_expire;
 pub mod text;
 pub mod timestamp;
